Add team ID to get retro,poker,storyboard db calls
