+++ internal/db/retro/retro.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"encoding/json"
-	"errors"
 	"fmt"
 
 	"github.com/StevenWeathers/thunderdome-planning-poker/internal/db"
@@ -27,6 +26,7 @@ func (d *Service) CreateRetro(ctx context.Context, OwnerID, TeamID string, Retro
 	var encryptedJoinCode string
 	var retro = &thunderdome.Retro{
 		OwnerID:               OwnerID,
+		TeamID:                TeamID,
 		Name:                  RetroName,
 		Phase:                 "intro",
 		PhaseTimeLimitMin:     PhaseTimeLimitMin,
@@ -163,7 +163,7 @@ func (d *Service) RetroGet(RetroID string, UserID string) (*thunderdome.Retro, e
 	var Template string
 	err := d.DB.QueryRow(
 		`SELECT
-			r.id, r.name, r.owner_id, r.phase, r.phase_time_limit_min, r.phase_time_start, r.phase_auto_advance,
+			r.id, r.name, r.owner_id, COALESCE(r.team_id::TEXT, ''), r.phase, r.phase_time_limit_min, r.phase_time_start, r.phase_auto_advance,
 			 COALESCE(r.join_code, ''), COALESCE(r.facilitator_code, ''), r.allow_cumulative_voting,
 			r.max_votes, r.brainstorm_visibility, r.ready_users, r.created_date, r.updated_date, r.template_id,
 			CASE WHEN COUNT(rf) = 0 THEN '[]'::json ELSE array_to_json(array_agg(rf.user_id)) END AS facilitators,
@@ -177,6 +177,7 @@ func (d *Service) RetroGet(RetroID string, UserID string) (*thunderdome.Retro, e
 		&b.Id,
 		&b.Name,
 		&b.OwnerID,
+		&b.TeamID,
 		&b.Phase,
 		&b.PhaseTimeLimitMin,
 		&b.PhaseTimeStart,
@@ -285,14 +286,14 @@ func (d *Service) RetroGetByUser(UserID string, Limit int, Offset int) ([]*thund
 		retros AS (
 			SELECT id from user_retros UNION SELECT id FROM team_retros
 		)
-		SELECT r.id, r.name, r.owner_id, r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.template_id,
+		SELECT r.id, r.name, r.owner_id, COALESCE(r.team_id::TEXT, ''), r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.template_id,
 		 r.allow_cumulative_voting, r.created_date, r.updated_date,
 		  MIN(COALESCE(t.name, '')) as teamName,
 		  (SELECT row_to_json(t.*) as template FROM thunderdome.retro_template t WHERE t.id = r.template_id) AS template
 		FROM thunderdome.retro r
 		LEFT JOIN user_teams t ON t.id = r.team_id
 		WHERE r.id IN (SELECT id FROM retros)
-		GROUP BY r.id ORDER BY r.created_date DESC LIMIT $2 OFFSET $3;
+		GROUP BY r.id, r.created_date ORDER BY r.created_date DESC LIMIT $2 OFFSET $3;
 	`, UserID, Limit, Offset)
 	if retrosErr != nil {
 		d.Logger.Error("get retros by user error", zap.Error(retrosErr))
@@ -309,6 +310,7 @@ func (d *Service) RetroGetByUser(UserID string, Limit int, Offset int) ([]*thund
 			&b.Id,
 			&b.Name,
 			&b.OwnerID,
+			&b.TeamID,
 			&b.Phase,
 			&b.PhaseTimeLimitMin,
 			&b.PhaseAutoAdvance,
@@ -332,169 +334,6 @@ func (d *Service) RetroGetByUser(UserID string, Limit int, Offset int) ([]*thund
 	}
 
 	return retros, Count, nil
-}
-
-// RetroConfirmFacilitator confirms the user is a facilitator of the retro
-func (d *Service) RetroConfirmFacilitator(RetroID string, userID string) error {
-	var facilitatorId string
-	var role string
-	err := d.DB.QueryRow("SELECT type FROM thunderdome.users WHERE id = $1", userID).Scan(&role)
-	if err != nil {
-		return fmt.Errorf("retro confirm facilitator get user role error: %v", err)
-	}
-
-	err = d.DB.QueryRow(
-		"SELECT user_id FROM thunderdome.retro_facilitator WHERE retro_id = $1 AND user_id = $2",
-		RetroID, userID).Scan(&facilitatorId)
-	if err != nil && role != thunderdome.AdminUserType {
-		return fmt.Errorf("get retro facilitator error: %v", err)
-	}
-
-	return nil
-}
-
-// RetroGetUsers retrieves the users for a given retro from db
-func (d *Service) RetroGetUsers(RetroID string) []*thunderdome.RetroUser {
-	var users = make([]*thunderdome.RetroUser, 0)
-	rows, err := d.DB.Query(
-		`SELECT
-			u.id, u.name, su.active, u.avatar, COALESCE(u.email, ''), COALESCE(u.picture, '')
-		FROM thunderdome.retro_user su
-		LEFT JOIN thunderdome.users u ON su.user_id = u.id
-		WHERE su.retro_id = $1
-		ORDER BY u.name;`,
-		RetroID,
-	)
-	if err == nil {
-		defer rows.Close()
-		for rows.Next() {
-			var w thunderdome.RetroUser
-			if err := rows.Scan(&w.ID, &w.Name, &w.Active, &w.Avatar, &w.Email, &w.PictureURL); err != nil {
-				d.Logger.Error("get retro users error", zap.Error(err))
-			} else {
-				if w.Email != "" {
-					w.GravatarHash = db.CreateGravatarHash(w.Email)
-				} else {
-					w.GravatarHash = db.CreateGravatarHash(w.ID)
-				}
-				users = append(users, &w)
-			}
-		}
-	}
-
-	return users
-}
-
-// GetRetroFacilitators gets a list of retro facilitator ids
-func (d *Service) GetRetroFacilitators(RetroID string) []string {
-	var facilitators = make([]string, 0)
-	rows, err := d.DB.Query(
-		`SELECT user_id FROM thunderdome.retro_facilitator WHERE retro_id = $1;`,
-		RetroID,
-	)
-	if err == nil {
-		defer rows.Close()
-		for rows.Next() {
-			var facilitator string
-			if err := rows.Scan(&facilitator); err != nil {
-				d.Logger.Error("get retro facilitators error", zap.Error(err))
-			} else {
-				facilitators = append(facilitators, facilitator)
-			}
-		}
-	}
-
-	return facilitators
-}
-
-// RetroAddUser adds a user by ID to the retro by ID
-func (d *Service) RetroAddUser(RetroID string, UserID string) ([]*thunderdome.RetroUser, error) {
-	if _, err := d.DB.Exec(
-		`INSERT INTO thunderdome.retro_user (retro_id, user_id, active)
-		VALUES ($1, $2, true)
-		ON CONFLICT (retro_id, user_id) DO UPDATE SET active = true, abandoned = false`,
-		RetroID,
-		UserID,
-	); err != nil {
-		d.Logger.Error("insert retro user error", zap.Error(err))
-	}
-
-	users := d.RetroGetUsers(RetroID)
-
-	return users, nil
-}
-
-// RetroFacilitatorAdd adds a retro facilitator
-func (d *Service) RetroFacilitatorAdd(RetroID string, UserID string) ([]string, error) {
-	if _, err := d.DB.Exec(
-		`INSERT INTO thunderdome.retro_facilitator (retro_id, user_id) VALUES ($1, $2);`,
-		RetroID, UserID); err != nil {
-		return nil, fmt.Errorf("retro add facilitator query error: %v", err)
-	}
-
-	facilitators := d.GetRetroFacilitators(RetroID)
-
-	return facilitators, nil
-}
-
-// RetroFacilitatorRemove removes a retro facilitator
-func (d *Service) RetroFacilitatorRemove(RetroID string, UserID string) ([]string, error) {
-	facilitatorCount := 0
-	err := d.DB.QueryRow(
-		`SELECT count(user_id) FROM thunderdome.retro_facilitator WHERE retro_id = $1;`,
-		RetroID,
-	).Scan(&facilitatorCount)
-	if err != nil {
-		return nil, fmt.Errorf("retro remove facilitator query error: %v", err)
-	}
-
-	if facilitatorCount == 1 {
-		return nil, fmt.Errorf("ONLY_FACILITATOR")
-	}
-
-	if _, err := d.DB.Exec(
-		`DELETE FROM thunderdome.retro_facilitator WHERE retro_id = $1 AND user_id = $2;`,
-		RetroID, UserID); err != nil {
-		return nil, fmt.Errorf("retro remove facilitator query error: %v", err)
-	}
-
-	facilitators := d.GetRetroFacilitators(RetroID)
-
-	return facilitators, nil
-}
-
-// RetroRetreatUser removes a user from the current retro by ID
-func (d *Service) RetroRetreatUser(RetroID string, UserID string) []*thunderdome.RetroUser {
-	if _, err := d.DB.Exec(
-		`UPDATE thunderdome.retro_user SET active = false WHERE retro_id = $1 AND user_id = $2`, RetroID, UserID); err != nil {
-		d.Logger.Error("update retro user active false error", zap.Error(err))
-	}
-
-	if _, err := d.DB.Exec(
-		`UPDATE thunderdome.users SET last_active = NOW() WHERE id = $1`, UserID); err != nil {
-		d.Logger.Error("update user last active timestamp error", zap.Error(err))
-	}
-
-	users := d.RetroGetUsers(RetroID)
-
-	return users
-}
-
-// RetroAbandon removes a user from the current retro by ID and sets abandoned true
-func (d *Service) RetroAbandon(RetroID string, UserID string) ([]*thunderdome.RetroUser, error) {
-	if _, err := d.DB.Exec(
-		`UPDATE thunderdome.retro_user SET active = false, abandoned = true WHERE retro_id = $1 AND user_id = $2`, RetroID, UserID); err != nil {
-		return nil, fmt.Errorf("abandon retro query error: %v", err)
-	}
-
-	if _, err := d.DB.Exec(
-		`UPDATE thunderdome.users SET last_active = NOW() WHERE id = $1`, UserID); err != nil {
-		return nil, fmt.Errorf("abandon retro update user query error: %v", err)
-	}
-
-	users := d.RetroGetUsers(RetroID)
-
-	return users, nil
 }
 
 // RetroAdvancePhase sets the phase for the retro
@@ -530,32 +369,6 @@ func (d *Service) RetroDelete(RetroID string) error {
 	return nil
 }
 
-// GetRetroUserActiveStatus checks retro active status of User for given retro
-func (d *Service) GetRetroUserActiveStatus(RetroID string, UserID string) error {
-	var active bool
-
-	err := d.DB.QueryRow(`
-		SELECT coalesce(active, FALSE)
-		FROM thunderdome.retro_user
-		WHERE user_id = $2 AND retro_id = $1;`,
-		RetroID,
-		UserID,
-	).Scan(
-		&active,
-	)
-	if err != nil && !errors.Is(err, sql.ErrNoRows) {
-		return fmt.Errorf("get retro user active status query error: %v", err)
-	} else if err != nil && errors.Is(err, sql.ErrNoRows) {
-		return err
-	}
-
-	if active {
-		return errors.New("DUPLICATE_RETRO_USER")
-	}
-
-	return nil
-}
-
 // GetRetros gets a list of retros
 func (d *Service) GetRetros(Limit int, Offset int) ([]*thunderdome.Retro, int, error) {
 	var retros = make([]*thunderdome.Retro, 0)
@@ -571,7 +384,7 @@ func (d *Service) GetRetros(Limit int, Offset int) ([]*thunderdome.Retro, int, e
 	}
 
 	rows, retrosErr := d.DB.Query(`
-		SELECT r.id, r.name, r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.allow_cumulative_voting,
+		SELECT r.id, COALESCE(r.team_id::TEXT, ''), r.name, r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.allow_cumulative_voting,
 		 r.created_date, r.updated_date, r.template_id,
 		 (SELECT row_to_json(t.*) as template FROM thunderdome.retro_template t WHERE t.id = r.template_id) AS template
 		FROM thunderdome.retro r
@@ -590,6 +403,7 @@ func (d *Service) GetRetros(Limit int, Offset int) ([]*thunderdome.Retro, int, e
 		var Template string
 		if err := rows.Scan(
 			&b.Id,
+			&b.TeamID,
 			&b.Name,
 			&b.Phase,
 			&b.PhaseTimeLimitMin,
@@ -630,7 +444,7 @@ func (d *Service) GetActiveRetros(Limit int, Offset int) ([]*thunderdome.Retro, 
 	}
 
 	rows, retrosErr := d.DB.Query(`
-		SELECT r.id, r.name, r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.allow_cumulative_voting, 
+		SELECT r.id, COALESCE(r.team_id::TEXT, ''), r.name, r.phase, r.phase_time_limit_min, r.phase_auto_advance, r.allow_cumulative_voting, 
 		r.created_date, r.updated_date,
 		r.template_id, (SELECT row_to_json(t.*) as template FROM thunderdome.retro_template t WHERE t.id = r.template_id) AS template
 		FROM thunderdome.retro_user ru
@@ -650,6 +464,7 @@ func (d *Service) GetActiveRetros(Limit int, Offset int) ([]*thunderdome.Retro, 
 		var Template string
 		if err := rows.Scan(
 			&b.Id,
+			&b.TeamID,
 			&b.Name,
 			&b.Phase,
 			&b.PhaseTimeLimitMin,
@@ -673,87 +488,4 @@ func (d *Service) GetActiveRetros(Limit int, Offset int) ([]*thunderdome.Retro, 
 	}
 
 	return retros, Count, nil
-}
-
-// GetRetroFacilitatorCode retrieve the retro facilitator code
-func (d *Service) GetRetroFacilitatorCode(RetroID string) (string, error) {
-	var EncryptedCode string
-
-	if err := d.DB.QueryRow(`
-		SELECT COALESCE(facilitator_code, '') FROM thunderdome.retro
-		WHERE id = $1`,
-		RetroID,
-	).Scan(&EncryptedCode); err != nil {
-		return "", fmt.Errorf("get retro facilitator_code query error: %v", err)
-	}
-
-	if EncryptedCode == "" {
-		return "", fmt.Errorf("retro facilitator_code not set")
-	}
-	DecryptedCode, codeErr := db.Decrypt(EncryptedCode, d.AESHashKey)
-	if codeErr != nil {
-		return "", fmt.Errorf("retrieve retro facilitator_code decrypt error: %v", codeErr)
-	}
-
-	return DecryptedCode, nil
-}
-
-// CleanRetros deletes retros older than {DaysOld} days
-func (d *Service) CleanRetros(ctx context.Context, DaysOld int) error {
-	if _, err := d.DB.ExecContext(ctx,
-		`DELETE FROM thunderdome.retro WHERE updated_date < (NOW() - $1 * interval '1 day');`,
-		DaysOld,
-	); err != nil {
-		return fmt.Errorf("clean retros query error: %v", err)
-	}
-
-	return nil
-}
-
-// MarkUserReady marks a user as ready for next phase
-func (d *Service) MarkUserReady(RetroID string, userID string) ([]string, error) {
-	var rawReadyUsers string
-	readyUsers := make([]string, 0)
-
-	err := d.DB.QueryRow(
-		`UPDATE thunderdome.retro 
-		SET updated_date = NOW(), ready_users = ready_users::jsonb || to_jsonb(array[$2])
-		WHERE id = $1 
-		RETURNING ready_users;`,
-		RetroID, userID,
-	).Scan(&rawReadyUsers)
-	if err != nil {
-		return readyUsers, fmt.Errorf("retro MarkUserReady query error: %v", err)
-	}
-
-	err = json.Unmarshal([]byte(rawReadyUsers), &readyUsers)
-	if err != nil {
-		d.Logger.Error("ready_users json error", zap.Error(err))
-	}
-
-	return readyUsers, nil
-}
-
-// UnmarkUserReady un-marks a user as ready for next phase
-func (d *Service) UnmarkUserReady(RetroID string, userID string) ([]string, error) {
-	var rawReadyUsers string
-	readyUsers := make([]string, 0)
-
-	err := d.DB.QueryRow(
-		`UPDATE thunderdome.retro 
-		SET updated_date = NOW(), ready_users = ready_users::jsonb - $2
-		WHERE id = $1 
-		RETURNING ready_users;`,
-		RetroID, userID,
-	).Scan(&rawReadyUsers)
-	if err != nil {
-		return readyUsers, fmt.Errorf("retro UnmarkUserReady query error: %v", err)
-	}
-
-	err = json.Unmarshal([]byte(rawReadyUsers), &readyUsers)
-	if err != nil {
-		d.Logger.Error("ready_users json error", zap.Error(err))
-	}
-
-	return readyUsers, nil
 }
