+++ thunderdome/storyboard.go
@@ -25,6 +25,7 @@ type Storyboard struct {
 	Personas        []*StoryboardPersona `json:"personas"`
 	JoinCode        string               `json:"joinCode" db:"join_code"`
 	FacilitatorCode string               `json:"facilitatorCode" db:"facilitator_code"`
+	TeamID          string               `json:"teamId" db:"team_id"`
 	TeamName        string               `json:"teamName"`
 	CreatedDate     string               `json:"createdDate" db:"created_date"`
 	UpdatedDate     string               `json:"updatedDate" db:"updated_date"`
